# api-gateway/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: 1.0.0
  description: Public API served through the API Gateway. Internal service endpoints are intentionally excluded.
servers:
  - url: /
security: []
...
paths:
...
  /otp/email/send:
    post:
      tags: [OTP]
      summary: Send email verification OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/email/verify:
    post:
      tags: [OTP]
      summary: Verify email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailOTPRequest"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/2fa/enable:
    post:
      tags: [OTP]
      summary: Send 2FA enable OTP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  maxLength: 128
                  format: password
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /otp/2fa/disable:
    post:
      tags: [OTP]
      summary: Send or verify 2FA disable OTP
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /otp/2fa/enable/verify:
    post:
      tags: [OTP]
      summary: Confirm 2FA enablement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/2fa/login/verify:
    post:
      tags: [OTP]
      summary: Complete login 2FA
      description: >
        Send the short-lived 2FA pending JWT returned by `/auth/login` in the
        Authorization header as `Bearer <2fa_pending_token>`. Normal access
        tokens are rejected.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: Access token and refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/password-reset/send:
    post:
      tags: [OTP]
      summary: Send password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/password-reset/verify:
    post:
      tags: [OTP]
      summary: Verify reset OTP and receive reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
  /otp/password-reset/apply:
    post:
      tags: [OTP]
      summary: Apply new password using reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reset_token, new_password, new_password_confirm]
              properties:
                reset_token:
                  type: string
                  maxLength: 256
                new_password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                new_password_confirm:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    OK:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
...
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Success:
      type: object
      required: [success, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorResponse:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
            retry_after_seconds:
              type: integer
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
    EmailOTPRequest:
      allOf:
        - $ref: "#/components/schemas/EmailRequest"
        - type: object
          required: [code]
          properties:
            code:
              type: string
              minLength: 6
              maxLength: 6
              pattern: "^[0-9]{6}$"
...